cli/requirements.txt
requests==2.31.0
click==8.1.3
tabulate==0.9.0
rich==13.3.5
pyjwt==2.6.0
python-dotenv==1.0.0
colorama==0.4.6
